requirements.txt
streamlit
SpeechRecognition
pyttsx3
pyaudio
Pillow
